chore: move some files out of server/common
This is a directory for sharing code between Node and Web.
