+++ packages/playwright-core/src/server/chromium/crConnection.ts
@@ -23,7 +23,7 @@ import { rewriteErrorMessage } from '../../utils/stackTrace';
 import { debugLogger, RecentLogsCollector } from '../../utils/debugLogger';
 import { ProtocolLogger } from '../types';
 import { helper } from '../helper';
-import { ProtocolError } from '../common/protocolError';
+import { ProtocolError } from '../protocolError';
 
 export const ConnectionEvents = {
   Disconnected: Symbol('ConnectionEvents.Disconnected')
