+++ packages/playwright-core/src/server/injected/componentUtils.ts
