+++ packages/playwright-core/src/server/protocolError.ts
