+++ tests/component-parser.spec.ts
@@ -15,8 +15,7 @@
  */
 
 import { playwrightTest as it, expect } from './config/browserTest';
-import type { ParsedComponentSelector } from '../packages/playwright-core/src/server/common/componentUtils';
-import { parseComponentSelector } from '../packages/playwright-core/lib/server/common/componentUtils';
+import { ParsedComponentSelector, parseComponentSelector } from '../packages/playwright-core/src/server/injected/componentUtils';
 
 const parse = parseComponentSelector;
 const serialize = (parsed: ParsedComponentSelector) => {
